api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
gunicorn==23.0.0
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
redis==5.2.1
pydantic-settings==2.8.1
PyJWT==2.10.1
pwdlib[argon2]==0.2.1
python-multipart==0.0.20
email-validator==2.2.0
